requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
pydantic==2.10.3
pydantic-settings==2.6.1
SQLAlchemy==2.0.36
alembic==1.14.0
asyncpg==0.30.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.19
httpx==0.27.2
pytest==8.3.3
pytest-asyncio==0.24.0
rich==13.9.4
orjson==3.10.12
greenlet>=3.0